@@ -6,7 +6,7 @@ description: Demonstrates how to use the flutter_pdf_split plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -22,10 +22,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
 
-  file_picker: ^2.1.0
-  permission_handler: ^5.0.1+1
+  file_picker: ^3.0.3
+  permission_handler: ^8.1.4+1
+  device_info_plus: ^2.1.0
 
 dev_dependencies:
   flutter_test:
