@@ -1,11 +1,11 @@
 name: flutter_pdf_split
 description: A new flutter plugin project.
-version: 0.0.1
+version: 0.2.0
 author: Mobizel
 homepage: https://github.com/lbaey/flutter_pdf_split
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
