@@ -12,8 +12,8 @@ class FlutterPdfSplitArgs {
   /// Prefix for each pdf page file, default value is 'page_'
   final String outFilePrefix;
 
-  FlutterPdfSplitArgs(this.filePath, this.outDirectory, {this.outFilePrefix = "page_"})
-      : assert(filePath != null && outDirectory != null);
+  FlutterPdfSplitArgs(this.filePath, this.outDirectory,
+      {this.outFilePrefix = "page_"});
 
   Map get toMap => {
         "filePath": filePath,
@@ -23,15 +23,15 @@ class FlutterPdfSplitArgs {
 }
 
 class FlutterPdfSplitResult {
-  int pageCount;
-  List<String> pagePaths;
+  int? pageCount;
+  late List<String> pagePaths;
 
   FlutterPdfSplitResult(Map result)
       : assert(result.containsKey("pageCount") &&
             result.containsKey("pagePaths") &&
             result["pagePaths"] is List) {
     this.pageCount = result["pageCount"];
-    this.pagePaths = List<String>();
+    this.pagePaths = <String>[];
     (result["pagePaths"] as List).forEach((path) {
       if (path is String) this.pagePaths.add(path);
     });
@@ -42,8 +42,8 @@ class FlutterPdfSplit {
   static const MethodChannel _channel =
       const MethodChannel('flutter_pdf_split');
 
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
+  static Future<String?> get platformVersion async {
+    final String? version = await _channel.invokeMethod('getPlatformVersion');
     return version;
   }
 
@@ -55,7 +55,7 @@ class FlutterPdfSplit {
   /// Splits PDF file [value] and returns the page count.
   static Future<FlutterPdfSplitResult> _split(FlutterPdfSplitArgs args) async {
     Map<dynamic, dynamic> result =
-        await _channel.invokeMethod('split', args.toMap);
+        await (_channel.invokeMethod('split', args.toMap));
     return FlutterPdfSplitResult(result);
   }
 }
